versions testing, +89 elo points
added an option to let two ai versions play, also removed the aspiration windows from the beginning of the game, made an improvement of 89 elo points.
possible bug: in the previous version there was a bug where the game ended when black still had moves. not sure if it's fixed or not yet, but i did fix several similar bugs, so it might be fixed.